requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.4.2
numpy==2.2.4
openpyxl==3.1.5
panda==0.3.1
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.3.0
matplotlib
seaborn
